Update README.md
Included the following sections: 

Added a section titled, ‘The Problem’ to define the target user and describe their pain points related to the challenge of writing tailored grant proposals.

Added a section titled, ‘Our Solution’ to describe our AI agent system as a collaborative team of specialist AI agents that automate burdensome administrative grant application tasks to an efficient flow of discovery, analysis, and drafting. 

Added a section titled, ‘The Value’ which defines the benefits of our application to provide an efficient grant application process.